add multer

+++ migrations/20191205064420-create-room.js
@@ -28,7 +28,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.FLOAT(7)
+        type: Sequelize.INTEGER(10)
       },
       createdAt: {
         allowNull: false,
